# src/lib/admin/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Tree-gateway",
        "description": "The Tree Gateway Administration API",
        "version": "1.0.0"
    },
    "host": "tree-gateway.com",
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "basePath": "/",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/apis/{apiId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve the API configuration for the given id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API Identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Api"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the API with the given id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API Identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "apiData",
                        "description": "API data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Api"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Invalid API",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API Identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "API cannot be deleted",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            }
        },
        "/apis/{apiId}/groups": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all groups for a specific API",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Group"
                            }
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new group",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Group data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Invalid group data",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            }
        },
        "/apis/{apiId}/groups/{groupId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a group by id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "groupId",
                        "description": "Group identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    },
                    "404": {
                        "description": "API/Group not found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a group",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "groupId",
                        "description": "Group identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Group data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Invalid group data",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API/Group not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a group",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "groupId",
                        "description": "Group identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cannot delete group",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API/Group not found"
                    }
                }
            }
        },
...
        "/apis/{apiId}/cache": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all cache configurations for the specified API",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Cache"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new cache configuration for the specified API",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Configuration data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Cache"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Invalid cache configuration",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            }
        },
        "/apis/{apiId}/cache/{cacheId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get the specified cache configuration.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "cacheId",
                        "description": "Cache ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Cache"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the specified cache configuration.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "cacheId",
                        "description": "Cache ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Configuration data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Cache"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cannot update cache configuration",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete the specified cache configuration.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "cacheId",
                        "description": "Cache ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cannot delete cache configuration",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Api": {
            "type": "object",
            "description": "API configuration",
            "properties": {
                "id": {
                    "type": "number",
                    "description": "The id unique identifier"
                },
                "name": {
                    "type": "string",
                    "description": "API name"
                },
                "description": {
                    "type": "string",
                    "description": "API description"
                },
                "url": {
                    "type": "string",
                    "description": "The URL for this API"
                }
            }
        },
        "Group": {
            "type": "object",
            "description": "Group data",
            "properties": {
                "id": {
                    "type": "number",
                    "description": "The id unique identifier"
                },
                "name": {
                    "type": "string",
                    "description": "Group name"
                }
            }
        },
...
        "Cache": {
            "type": "object",
            "description": "Configuration for API cache.",
            "properties": {
                "client": {
                    "$ref": "#/definitions/ClientCache"
                },
                "server": {
                    "$ref": "#/definitions/ServerCache"
                },
                "groups": {
                    "description": "List of groups that should be handled by this limiter. If not provided, everything will be handled.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "ClientCache": {
            "type": "object",
            "description": "Configuration for a client side cache (in browser).",
            "properties": {
                "cacheTime": {
                    "type": "string"
                },
                "cacheControl": {
                    "type": "string"
                },
                "mustRevalidate": {
                    "type": "boolean"
                },
                "noTransform": {
                    "type": "boolean"
                },
                "proxyRevalidate": {
                    "type": "boolean"
                }
            }
        },
        "ServerCache": {
            "type": "object",
            "description": "Configuration for a server side cache (in Memory or with a Redis store)",
            "properties": {
                "cacheTime": {
                    "type": "string"
                },
                "binary": {
                    "type": "boolean"
                },
                "preserveHeaders": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32",
                    "description": "Error code."
                },
                "message": {
                    "type": "string",
                    "description": "Error message"
                }
            }
        }
    }
}